Move PackedSolitaire into its own file
