+++ Assets/Scripts/Core/PackedSolitaire.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 99e06ff2b156ef948b6ebd21c9f288a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
